Add files via upload
The BMI (Body Mass Index) calculator is a simple tool used to assess an individual's body weight relative to their height. By inputting your height and weight, the BMI calculator provides a numerical value that falls into specific categories, such as underweight, normal weight, overweight, or obese. This measurement is widely used as an initial indicator of potential health risks associated with body weight and can be a helpful tool for individuals striving for a healthier lifestyle. In those files i create some different html files like 11-20.html, 21-30.html so this files are help users to understand their BMI Value and and they get a proper healthy diet plan for their health. However, it is essential to remember that BMI is a general estimation and does not take into account factors such as muscle mass or distribution of weight, so it should be considered alongside other health assessments for a comprehensive evaluation.